# car-models-service-core/docs/api-docs.yml
openapi: 3.0.1
info:
  title: Car models service API
  description: The API to persist and search car models in a database
  version: "1.0"
servers:
- url: http://localhost:8180
  description: Generated server url
paths:
  /v1/manufacturers/{manufacturer}/models/{name}/{year}:
    get:
      tags:
      - model
      summary: "Get a model by its manufacturer, name and year"
      description: "Seach for and retrieve a model by its manufacturer, name and year\
        \ from the database"
      operationId: getModelByManufacturerAndNameAndYear
      parameters:
      - name: manufacturer
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: year
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelDto'
              examples:
                model:
                  description: model
                  value:
                    id: "1"
                    name: A7
                    year: "2023"
                    manufacturer: Audi
                    categories:
                    - SUV
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
    put:
      tags:
      - model
      summary: Update a model
      description: "Seach for a model by its manufacturer, name and year in the database\
        \ if it exists, update it"
      operationId: updateModel
      parameters:
      - name: manufacturer
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: year
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelDto'
            examples:
              model:
                description: model
                value:
                  id: null
                  name: null
                  year: null
                  manufacturer: null
                  categories:
                  - Sedan
        required: true
      responses:
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Ok
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
    post:
      tags:
      - model
      summary: Save a model
      description: "Seach for a model by its manufacturer, name and year in the database\
        \ if it is missing, persist it"
      operationId: saveModel
      parameters:
      - name: manufacturer
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: year
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelDto'
            examples:
              model:
                description: model
                value:
                  id: null
                  name: null
                  year: null
                  manufacturer: null
                  categories:
                  - Sedan
        required: true
      responses:
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "201":
          description: Created
          headers:
            Location:
              description: "/v1/models/{id}"
              style: simple
              schema:
                type: string
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
...
  /v1/models:
    get:
      tags:
      - model
      summary: Search models
      description: "Seach for models in the database by optional parameters if no\
        \ one is specified, retrieve all models"
      operationId: searchModels
      parameters:
      - name: model
        in: query
        required: false
        schema:
          type: string
      - name: category
        in: query
        required: false
        schema:
          type: string
      - name: manufacturer
        in: query
        required: false
        schema:
          type: string
      - name: maxYear
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: minYear
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: year
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 100
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageModelDto'
      security:
      - bearerAuth: []
  /v1/models/{id}:
    get:
      tags:
      - model
      summary: Get a model by its id
      description: Seach for and revrieve a model from the database by its name
      operationId: getModelById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelDto'
              examples:
                model:
                  description: model
                  value:
                    id: "1"
                    name: A7
                    year: "2023"
                    manufacturer: Audi
                    categories:
                    - Sedan
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
    delete:
      tags:
      - model
      summary: Delete a model by its id
      description: Search for and delete from the database a model by its id
      operationId: deleteModelById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "405":
          description: Method Not Allowed
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "204":
          description: No Content
      security:
      - bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
    Violation:
      type: object
      properties:
        attribute:
          type: string
        message:
          type: string
    ModelDto:
      required:
      - categories
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        year:
          type: integer
          format: int32
        manufacturer:
          type: string
        categories:
          uniqueItems: true
          type: array
          items:
            type: string
...
    PageModelDto:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/ModelDto'
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        empty:
          type: boolean
        unsorted:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT